+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file. `Jayme` a
 - `ServerBackendConfiguration` has been renamed to `NSURLSessionBackendConfiguration`.
 - `ServerBackendError` has been renamed to `JaymeError`.
 - `ServerRepository` has been renamed to `CRUDRepository`.
+- `StringDictionary` typealias has been removed.
 
 ---